# backend/openapi.yaml
openapi: 3.0.0
paths:
...
  /users:
    get:
      operationId: UsersController_list
      parameters:
        - name: search
          required: false
          in: query
          description: Free text search by name or email
          schema:
            type: string
        - name: role
          required: false
          in: query
          schema:
            type: string
            enum:
              - USER
              - ADMIN
      responses:
        '200':
          description: Returns a paginated list of users.
      security: &ref_1
        - bearer: []
      summary: List users
      tags: &ref_2
        - Users
    post:
      operationId: UsersController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '201':
          description: User created successfully.
      security: *ref_1
      summary: Create a user
      tags: *ref_2
  /users/{id}:
    get:
      operationId: UsersController_detail
      parameters:
        - name: id
          required: true
          in: path
          description: User identifier
          schema:
            type: number
      responses:
        '200':
          description: Returns a single user.
      security: *ref_1
      summary: Get user by id
      tags: *ref_2
    patch:
      operationId: UsersController_update
      parameters:
        - name: id
          required: true
          in: path
          description: User identifier
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
      responses:
        '200':
          description: User updated successfully.
      security: *ref_1
      summary: Update a user
      tags: *ref_2
    delete:
      operationId: UsersController_remove
      parameters:
        - name: id
          required: true
          in: path
          description: User identifier
          schema:
            type: number
      responses:
        '200':
          description: User deleted successfully.
      security: *ref_1
      summary: Delete a user
      tags: *ref_2
...
  /tasks:
    get:
      operationId: TasksController_list
      parameters:
        - name: status
          required: false
          in: query
          schema:
            type: array
            items:
              type: string
              enum:
                - TODO
                - IN_PROGRESS
                - COMPLETED
                - ARCHIVED
        - name: priority
          required: false
          in: query
          schema:
            type: array
            items:
              type: string
              enum:
                - LOW
                - MEDIUM
                - HIGH
                - URGENT
        - name: dueFrom
          required: false
          in: query
          schema:
            format: date-time
            type: string
        - name: dueTo
          required: false
          in: query
          schema:
            format: date-time
            type: string
        - name: tag
          required: false
          in: query
          description: Filter by a single tag name
          schema:
            type: string
        - name: tags
          required: false
          in: query
          description: Filter by one or more tag names
          schema:
            type: array
            items:
              type: string
        - name: q
          required: false
          in: query
          description: Full text search on title or description
          schema:
            type: string
        - name: includeSubtasks
          required: false
          in: query
          description: Include subtasks in top-level listing
          schema:
            default: false
            type: boolean
      responses:
        '200':
          description: Returns a paginated list of tasks that match the provided filters.
      security: &ref_5
        - bearer: []
      summary: List tasks
      tags: &ref_6
        - Tasks
    post:
      operationId: TasksController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskDto'
      responses:
        '201':
          description: Task created successfully.
      security: *ref_5
      summary: Create a new task
      tags: *ref_6
  /tasks/{id}:
    get:
      operationId: TasksController_detail
      parameters:
        - name: id
          required: true
          in: path
          description: Task identifier
          schema:
            type: number
      responses:
        '200':
          description: Returns a single task.
      security: *ref_5
      summary: Get task by id
      tags: *ref_6
    patch:
      operationId: TasksController_update
      parameters:
        - name: id
          required: true
          in: path
          description: Task identifier
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskDto'
      responses:
        '200':
          description: Task updated successfully.
      security: *ref_5
      summary: Update an existing task
      tags: *ref_6
    delete:
      operationId: TasksController_remove
      parameters:
        - name: id
          required: true
          in: path
          description: Task identifier
          schema:
            type: number
      responses:
        '200':
          description: Task deleted successfully.
      security: *ref_5
      summary: Delete a task
      tags: *ref_6
  /tasks/{id}/tags/{tagId}:
    post:
      operationId: TasksController_attachTag
      parameters:
        - name: id
          required: true
          in: path
          description: Task identifier
          schema:
            type: number
        - name: tagId
          required: true
          in: path
          description: Tag identifier
          schema:
            type: number
      responses:
        '200':
          description: Updated task with the tag attached.
      security: *ref_5
      summary: Assign a tag to a task
      tags: *ref_6
    delete:
      operationId: TasksController_detachTag
      parameters:
        - name: id
          required: true
          in: path
          description: Task identifier
          schema:
            type: number
        - name: tagId
          required: true
          in: path
          description: Tag identifier
          schema:
            type: number
      responses:
        '200':
          description: Updated task with the tag removed.
      security: *ref_5
      summary: Remove a tag from a task
      tags: *ref_6
  /tasks/{id}/comments:
    get:
      operationId: TasksController_comments
      parameters:
        - name: id
          required: true
          in: path
          description: Task identifier
          schema:
            type: number
      responses:
        '200':
          description: Returns comments for the given task.
      security: *ref_5
      summary: List comments for a task
      tags: *ref_6
    post:
      operationId: TasksController_addComment
      parameters:
        - name: id
          required: true
          in: path
          description: Task identifier
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentDto'
      responses:
        '201':
          description: Comment created successfully.
      security: *ref_5
      summary: Create a comment on a task
      tags: *ref_6
  /tasks/{id}/subtasks:
    post:
      operationId: TasksController_addSubtask
      parameters:
        - name: id
          required: true
          in: path
          description: Parent task identifier
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskDto'
      responses:
        '201':
          description: Subtask created successfully.
      security: *ref_5
      summary: Create a subtask
      tags: *ref_6
...
info:
  title: Finace API
  description: OpenAPI documentation for the Finace backend services.
  version: 1.0.0
  contact: {}
...
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    jwt:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: Provide the access token obtained from the auth endpoints.
  schemas:
...
    CreateUserDto:
      type: object
      properties:
        email:
          type: string
          example: user@example.com
        password:
          type: string
          example: StrongPassword1!
          minLength: 8
        name:
          type: string
          example: Jane Doe
        role:
          type: string
          enum:
            - USER
            - ADMIN
          example: USER
      required:
        - email
        - password
        - name
    UpdateUserDto:
      type: object
      properties:
        email:
          type: string
          example: user@example.com
        password:
          type: string
          example: StrongPassword1!
          minLength: 8
        name:
          type: string
          example: Jane Doe
        role:
          type: string
          enum:
            - USER
            - ADMIN
          example: USER
...
    CreateTaskDto:
      type: object
      properties:
        title:
          type: string
          example: Review quarterly budget
        description:
          type: string
          example: Ensure all expense reports are collected
          maxLength: 5000
        status:
          type: string
          enum:
            - TODO
            - IN_PROGRESS
            - COMPLETED
            - ARCHIVED
          default: TODO
        priority:
          type: string
          enum:
            - LOW
            - MEDIUM
            - HIGH
            - URGENT
          default: MEDIUM
        dueDate:
          type: string
          format: date-time
        repeatRRule:
          type: string
          description: RFC5545 RRULE definition for repeating tasks
        tagNames:
          example:
            - finance
            - reporting
          type: array
          items:
            type: string
        parentId:
          type: number
          description: Parent task identifier for subtasks
      required:
        - title
    UpdateTaskDto:
      type: object
      properties:
        title:
          type: string
          example: Review quarterly budget
        description:
          type: string
          example: Ensure all expense reports are collected
          maxLength: 5000
        status:
          type: string
          enum:
            - TODO
            - IN_PROGRESS
            - COMPLETED
            - ARCHIVED
        priority:
          type: string
          enum:
            - LOW
            - MEDIUM
            - HIGH
            - URGENT
        dueDate:
          type: string
          format: date-time
        repeatRRule:
          type: object
          description: RFC5545 RRULE definition for repeating tasks
        tagNames:
          example:
            - finance
            - reporting
          type: array
          items:
            type: string
        parentId:
          type: object
          description: Parent task identifier or null to detach
          nullable: true
    CreateCommentDto:
      type: object
      properties:
        text:
          type: string
          example: Let us sync before Friday.
      required:
        - text
...